Use 7 dots, 4 colors, each colour appearing just once
This will potentially be more robust as we can (as
a next step) implement local calibration per corner.

Also 4 colors instead of 5 should be more robust.

The 7 dots will also give us more combinations
(2100 papers instead of 625), and it’s probably ok
to scale down the size of the dots later.